Add support for Flexmojos 7.1
