--- net.flexmojos.m2e.test/META-INF/MANIFEST.MF
@@ -1,8 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Test
-Bundle-SymbolicName: net.flexmojos.m2e.test
-Bundle-Version: 1.0.0.qualifier
-Fragment-Host: net.flexmojos.m2e;bundle-version="1.0.0"
-Bundle-RequiredExecutionEnvironment: JavaSE-1.6
-Require-Bundle: org.junit;bundle-version="4.8.2"
