--- net.flexmojos.m2e.test/build.properties
@@ -1,4 +1,0 @@
-source.. = src/
-output.. = target/classes
-bin.includes = META-INF/,\
-               .
