--- net.flexmojos.m2e.test/pom.xml
@@ -1,14 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	
-	<parent>
-		<groupId>net.flexmojos.oss</groupId>
-		<artifactId>m2e-flexmojos</artifactId>
-		<version>1.0.0-SNAPSHOT</version>
-	</parent>
-	
-	<artifactId>net.flexmojos.m2e.test</artifactId>
-	<packaging>eclipse-test-plugin</packaging>
-	
-	<name>Flexmojos Integration for Flash Builder Test Plug-in</name>
-</project>
