+++ net.flexmojos.m2e/src/net/flexmojos/m2e/maven/internal/fm7/SignAirMojo.java
@@ -1,4 +1,4 @@
-package net.flexmojos.m2e.maven.internal.fm6;
+package net.flexmojos.m2e.maven.internal.fm7;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.METHOD;
@@ -11,7 +11,7 @@ import java.lang.annotation.Target;
 import com.google.inject.BindingAnnotation;
 
 @BindingAnnotation @Target({ FIELD, PARAMETER, METHOD }) @Retention(RUNTIME)
-public @interface CompilerMojo
+public @interface SignAirMojo
 {
 
 }
