+++ net.flexmojos.m2e/src/net/flexmojos/m2e/maven/internal/fm7/adapters/Flexmojos7SignAirMojo.java
@@ -1,8 +1,8 @@
-package net.flexmojos.m2e.maven.internal.fm6.adapters;
+package net.flexmojos.m2e.maven.internal.fm7.adapters;
 
 import net.flexmojos.m2e.maven.ISignAirMojo;
 import net.flexmojos.m2e.maven.internal.MavenFlexMojo;
-import net.flexmojos.m2e.maven.internal.fm6.SignAirMojo;
+import net.flexmojos.m2e.maven.internal.fm7.SignAirMojo;
 
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.MojoExecution;
@@ -10,10 +10,10 @@ import org.eclipse.jdt.annotation.Nullable;
 
 import com.google.inject.Inject;
 
-public class Flexmojos6SignAirMojo extends MavenFlexMojo implements ISignAirMojo
+public class Flexmojos7SignAirMojo extends MavenFlexMojo implements ISignAirMojo
 {
 
-    @Inject Flexmojos6SignAirMojo( final MavenSession session,
+    @Inject Flexmojos7SignAirMojo( final MavenSession session,
                                    final @SignAirMojo @Nullable MojoExecution mojoExecution )
     {
         super( session, mojoExecution );
